Successfully adds Register(), a function component to validate login in credentials, broke css
